Highlight current apparatus entry.
We're adding the class 'current' to the innermost .appnote element that
is hovered over. This coincedes with the one of which we take the title
element.

Now we need to cleanup the apparatus encoding such that there is one
span surrounding each apparatus stuff and the underlines and generated
text stuff immediately below that.