Adjusted testimony processing to copy all table entries.
get_testimonies.py will now write all columns from the excel table to
xslt/testimony-table.xml, thus enabling to work with fields that are not
displayed in the table.

In order to reduce the amount of data, the XML file will only contain
elements for fields that have values for the corresponding testimony.
Additionaly, the XML file contains a header with an entry for each
column of the original Excel table.

To reduce errors from strange table headers, each table header from the
excel table is 'slugified':

* unicode-normalized
* converted to lower case
* äöüß expanded
* stretches of non-identifier-chars replaced with -
* leading/trailing - stripped

So " Biedermann-HerwigNr." becomes "biedermann-herwignr" and
"Datum.(von)" becomes "datum-von". This is also reflected in the XML
header.

The XSLT stylesheet now contains the config for the columns and column
labels to be used, so there is hopefully a single place for everything
related to representation.