resolve-uri replacement that accepts empty static base.
fn:static-base-uri() returns the empty sequence inside the XsltServlet
in eXist-db ≥ 5.0, and we sometimes use resolve-uri in situations where
this is not permitted although it wouldn’t be a problem. This caused
those searches that render their result through the default HTML
pipeline to fail. This commit uses a safe implementation that just
returns a relative URI instead in these situations.

Fixes #648.