Links to the watermark image table.
If we have a watermark image, display an image icon behind the
respective watermark’s label that is a link into the table with the
watermark images. In the case of the table, the whole label can be
a link, in case of the individual witnesses’ metadata files, the label
links into the table of all watermarks.